+++ R/function_PlotMapPoints.R
@@ -105,7 +105,7 @@
 
 #' @examples
 #' \donttest{
-#' # Import plot data and subbasin polygons
+#' # Import plot data and subbasin points
 #' require(sf)
 #' te1 <- ReadSubass(filename = system.file("demo_model",
 #' "results", "subass1.txt", package = "HYPEtools"))