+++ NEWS.md
@@ -1,10 +1,10 @@
 # HYPEtools NEWS
 
-## HYPEtools 1.5.0.9000
+## HYPEtools 1.5.1
 *Development Version*
 
 #### Highlights
-* Fix broken @docTYPE package documentation as described here: https://github.com/r-lib/roxygen2/issues/1491
+* Fix broken @docTYPE package documentation (package?pkgname) as described here: https://github.com/r-lib/roxygen2/issues/1491
 * Update PlotMapOutput() and PlotMapPoints() to allow for missing SUBIDs in MapOutput or GIS files
 * Update PlotMapOutput() and PlotMapPoints() functions so "map"/"sites" and "map.subid.column"/"sites.subid.column" arguments are interchangeable